Michael 5 user authentication (#7)
* User Authentication.
Searches through database via repository/UserRepository
User Service returns a user if it is found.  If not, it creates a new user object.
User Service can also return if it exists or not.
Message will be shown in welcome.html if the user exists.  If they are a new user, a message will show that a new user was detected.

* Removed application.properties

* User Changes for autoincrement.
Changed long to int for id
Mysql changed to auto_increment